refactor: split out merge execute
